Convert Progress to TypeScript
